% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_fun.R
\name{backup_project}
\alias{backup_project}
\title{Export the R project folder as a compressed file}
\usage{
backup_project(
  overwrite = FALSE,
  prefix = NULL,
  timestamp = FALSE,
  output_path = NULL,
  include_data = FALSE,
  ignore = NULL,
  include = NULL,
  ...
)
}
\arguments{
\item{overwrite}{Overwrite target file or not. Is bypassed if timestamp is TRUE.}

\item{prefix}{What the exported project file should use as prefix. Defaults to 'project name' using basename()}

\item{timestamp}{If TRUE, adds a timestamp in addition to a date-suffix.}

\item{output_path}{Defaults to ./backup/}

\item{include_data}{Should subdirectory of ./data/ be included in export?}

\item{ignore}{Additional files to ignore (Currently not implemented!)}

\item{include}{Additional files to include (Currently not implemented!)}

\item{...}{Additional arguments that can be passed down}
}
\description{
\code{backup_project()}exports the R project folder for easy backup and/or sharing local versions
}
\details{
It exports files present in the working directory of the current R project.
It is flexible in terms of output location and white- and blacklisting of files to include.
Due note, by default it does not export any files that are present in ./data/ in case
of any sensitive data present. However, this is available to bypass as an argument.
}
