% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{extract_libs}
\alias{extract_libs}
\title{Function to extract package names from a pre-determined set of data}
\usage{
extract_libs(lib)
}
\arguments{
\item{lib}{List of package names}
}
\description{
Function to extract package names from a pre-determined set of data
}
