% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_test}
\alias{generate_test}
\title{Generate a test report}
\usage{
generate_test(...)
}
\arguments{
\item{...}{Optional arguments for parent function}
}
\description{
Generate a test report
}
