% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libs.R
\name{get_libs}
\alias{get_libs}
\title{Install from a list of packages}
\usage{
get_libs(lib = "all")
}
\arguments{
\item{lib}{What group of packages is wanted. Default is "all".
base - Main packages used by this package
epi - Packages for epidemiological data handling
markdown - Packages for different type of markdown handling
bench - Packages for benchmarking purposes
dev - Packages for development purposes}
}
\description{
\code{get_libs()} checks for existing packages and ask for permission if any are missing.
}
