% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{handle_libs}
\alias{handle_libs}
\title{Function to handle input and either load or unload packages}
\usage{
handle_libs(lib, method)
}
\description{
Function to handle input and either load or unload packages
}
