% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{install_libs}
\alias{install_libs}
\title{Install from a list of packages}
\usage{
install_libs(lib = "all", internal = TRUE)
}
\arguments{
\item{lib}{Default 'all'; Vector of either package names or use internal sets}

\item{internal}{Default 'TRUE'; Whether to use internal pre-determined sets}
}
\description{
\code{install_libs()} checks for packages (user input or pre-determined packages;
see below) and ask for permission to install if any are missing.
}
\details{
"base" - Main packages used by this package

"epi" - Packages for epidemiological data handling

"mdown" - Packages for different type of markdown handling

"bench" - Packages for benchmarking purposes

"dev" - Packages for development purposes
}
