% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{load_libs}
\alias{load_libs}
\title{Load packages from a pre-determined set.}
\usage{
load_libs(lib)
}
\arguments{
\item{lib}{Default NULL; See \code{install_libs()} for which sets are available.}
}
\description{
Load packages from a pre-determined set.
}
