% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{print_libs}
\alias{print_libs}
\title{Retrieves and prints what the packages are in the pre-determined sets as a character vector}
\usage{
print_libs(group = "all")
}
\arguments{
\item{group}{Default 'all'; Group name of pre-determined sets of packages}
}
\description{
Retrieves and prints what the packages are in the pre-determined sets as a character vector
}
