% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fun.R
\name{print_rmd}
\alias{print_rmd}
\title{Helper function for easier printing of Rmd files}
\usage{
print_rmd(
  name,
  params = NULL,
  output_path = "./output",
  open = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{name}{Name of Rmd file that will be used. Locates in './Rmd/'.}

\item{params}{Pass along parameters/data to the Rmd}

\item{output_path}{Defaults './output'; Where to output rendered Rmd files}

\item{open}{Logical; Open the rendered file when done?}

\item{quiet}{Logical; Should rendering be quiet?}
}
\description{
\code{print_rmd()}returns labels from a data dictionary, either renaming the columns or
returns a character vector with the labels. The latter may be useful when wanting
to label variables used in plots or tables.
}
