% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{remove_libs}
\alias{remove_libs}
\title{Uninstalls packages from a pre-determined set}
\usage{
remove_libs(lib)
}
\arguments{
\item{lib}{Default NULL; Must explicitly state which set of packages.

See \code{install_libs()} for further information.}
}
\description{
Uninstalls packages from a pre-determined set
}
