% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roosdorp-package.R
\docType{package}
\name{roosdorp-package}
\alias{roosdorp}
\alias{roosdorp-package}
\title{roosdorp: Personal R Package With Misc Functions}
\description{
A compilation of miscellaneous functions to help with uncommon tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phiido/roosdorp}
  \item Report bugs at \url{https://github.com/phiido/roosdorp/issues}
}

}
\author{
\strong{Maintainer}: Thomas Roosdorp \email{github@roosdorp.se}

}
\keyword{internal}
