% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libs_fun.R
\name{unload_libs}
\alias{unload_libs}
\title{Unloads packages from a pre-determined set.}
\usage{
unload_libs(lib)
}
\arguments{
\item{lib}{Default NULL; See \code{install_libs()} for which sets are available.}
}
\description{
Unloads packages from a pre-determined set.
}
