% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Render custom reports from data.frame}
\usage{
generate_report(
  data,
  var,
  template,
  by = NULL,
  overwrite = TRUE,
  output_folder = "./output/reports",
  ...
)
}
\arguments{
\item{data}{Input as data frame.}

\item{var}{Name of variable containing UID for reports.}

\item{template}{File path with Rmd-template for rendering.}

\item{by}{Variable to group reports into sub-folders. Default is NULL.}

\item{overwrite}{Whether existing files should be overwritten.}

\item{output_folder}{Folder to use for output. Default (relative to project): 'output/reports'.}

\item{...}{Additional arguments than can be passed down to
rmarkdown:render(). Argument quiet is set to TRUE.}
}
\description{
\code{generate_report()}renders multiple reports from custom Rmd-template, and
uses parallel computing if possible for faster processing.
}
\details{
Make sure there is sufficient memory and CPU available. Rmarkdown/tinytex may
throw errors otherwise. Mostly an issue if running on a laptop.
}
