% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Render reports}
\usage{
generate_report(df, var, template, ...)
}
\arguments{
\item{df}{Input as data frame}

\item{var}{Name of variable containing UID for reports}

\item{template}{File path with Rmd-template for rendering

Optional arguments}

\item{by}{Which variable should be used for grouping outputs, or none}

\item{overwrite}{Whether existing files should be overwritten}

\item{is_quiet}{Whether Rmarkdown should be quiet}

\item{output_folder}{Folder to use for output. Default: './output/reports'}
}
\description{
Render multiple reports from Rmd-template. Can use parallel computing if possible.
Make sure there is sufficient memory and CPU available. Rmarkdown/tinytex may throw errors otherwise.
}
\author{
Thomas Roosdorp
}
