% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun.R
\name{label_vars}
\alias{label_vars}
\title{Renames variable names based on data dictionary}
\usage{
label_vars(data, dict = NULL, simple = FALSE)
}
\arguments{
\item{data}{Input as data frame or character.vector}

\item{dict}{Input as data.frame with two columns. Should closely follow a key value structure}

\item{simple}{Logical; Should return as character.vector? Otherwise renamed data.frame}
}
\value{
Data.frame or character.vector
}
\description{
\code{label_vars()}returns labels from a data dictionary, either renaming the columns or
returns a character vector with the labels. The latter may be useful when wanting
to label variables used in plots or tables.
}
